ssfaes: fix signed-shift UB in _SSFAESKeyExpansion's byte-load loop.
UBSan (gcc 15 -fsanitize=undefined on the unit suite) flagged
_crypto/ssfaes.c:218 — `key[t+3] << 24` triggers UB whenever the
key byte has its high bit set:

  w[i] = (key[t]) ^ (key[t + 1] << 8) ^ (key[t + 2] << 16)
       ^ (key[t + 3] << 24);

`key[t]` is `uint8_t`; the integer-promotion rules promote it to
`int` before the shift. Shifting a value >= 0x80 left by 24 produces
a result >= 0x80000000, which exceeds INT_MAX on every supported
target — undefined behavior per C11 6.5.7 paragraph 4.

In practice every compiler we've tested produces the expected
2's-complement codegen and the AES key schedule is correct. But this
is exactly the class of bug that was *just* fixed in
ARRAY_TO_STATE / STATE_TO_ARRAY (commit 0119f48 "ssfaes: fix big-
endian KAT failures...") via the same uint32_t-cast pattern — a
future toolchain version, a different optimizer pass, or a stricter
-fsanitize=undefined CI gate could turn this into a real
miscompile.

Fix: cast each byte to uint32_t before shifting, matching the
ARRAY_TO_STATE idiom. Equivalent codegen on every compiler the
cross matrix exercises (mipsel, m68k, powerpc, etc.); strictly-
defined per the standard.

Verified: full unit suite under gcc-15 -fsanitize=address,undefined
on Linux x86_64 reports zero findings after the fix. Wycheproof
ECDSA P-256 (482), AES-CTR NIST F.5, and the AES Monte Carlo KATs
all still pass on macOS / linux / 12-arch QEMU cross matrix.

+++ _crypto/ssfaes.c
@@ -215,7 +215,8 @@ static void _SSFAESKeyExpansion(uint32_t *w, size_t wSize, const uint8_t *key, s
     for (i = 0; i < nk; i++)
     {
         t = i << 2;
-        w[i] = (key[t]) ^ (key[t + 1] << 8) ^ (key[t + 2] << 16) ^ (key[t + 3] << 24);
+        w[i] = ((uint32_t)key[t])              ^ ((uint32_t)key[t + 1] <<  8) ^
+               ((uint32_t)key[t + 2] << 16)    ^ ((uint32_t)key[t + 3] << 24);
     }
 
     for (i = nk; i < wSize; i++)
